+++ buildSrc/src/main/kotlin/org/jetbrains/exposed/gradle/Versions.kt
@@ -9,18 +9,19 @@ object Versions {
 
     /** JDBC drivers **/
     const val h2 = "1.4.199"
-    const val mariaDB = "2.7.0"
+    const val h2_v2 = "2.0.202"
+    const val mariaDB = "2.7.4"
     const val mysql51 = "5.1.49"
-    const val mysql80 = "8.0.23"
+    const val mysql80 = "8.0.27"
     const val oracle12 = "12.2.0.1"
-    const val postgre = "42.2.19"
-    const val postgreNG = "0.8.6"
-    const val sqlLite3 = "3.36.0.1"
+    const val postgre = "42.3.1"
+    const val postgreNG = "0.8.9"
+    const val sqlLite3 = "3.36.0.3"
     const val sqlserver = "8.4.1.jre8"
 
     /** Spring **/
-    const val springFramework = "5.3.7"
-    const val springBoot = "2.5.0"
+    const val springFramework = "5.3.13"
+    const val springBoot = "2.6.1"
 
     /** Test Dependencies **/
     const val testContainers = "1.16.2"