+++ exposed-core/src/main/kotlin/org/jetbrains/exposed/sql/AbstractQuery.kt
@@ -4,7 +4,6 @@ import org.jetbrains.exposed.sql.statements.Statement
 import org.jetbrains.exposed.sql.statements.StatementType
 import org.jetbrains.exposed.sql.transactions.TransactionManager
 import java.sql.ResultSet
-import java.util.concurrent.ConcurrentHashMap
 
 abstract class AbstractQuery<T : AbstractQuery<T>>(targets: List<Table>) : SizedIterable<ResultRow>, Statement<ResultSet>(StatementType.SELECT, targets) {
     protected val transaction get() = TransactionManager.current()
@@ -60,49 +59,50 @@ abstract class AbstractQuery<T : AbstractQuery<T>>(targets: List<Table>) : Sized
 
     override fun iterator(): Iterator<ResultRow> {
         val resultIterator = ResultIterator(transaction.exec(queryToExecute)!!)
-        return if (transaction.db.supportsMultipleResultSets) resultIterator.also { trackStatementOpen(transaction, it) }
-        else {
+        return if (transaction.db.supportsMultipleResultSets) {
+            resultIterator
+        } else {
             Iterable { resultIterator }.toList().iterator()
         }
     }
 
-    protected inner class ResultIterator(val rs: ResultSet) : Iterator<ResultRow> {
-        private var hasNext: Boolean? = null
+    private inner class ResultIterator(val rs: ResultSet) : Iterator<ResultRow> {
+        private var hasNext = false
+            set(value) {
+                field = value
+                if (!field) {
+                    rs.statement?.close()
+                    transaction.openResultSetsCount--
+                }
+            }
 
         private val fieldsIndex = set.realFields.toSet().mapIndexed { index, expression -> expression to index }.toMap()
 
-        override operator fun next(): ResultRow {
-            if (hasNext == null) hasNext()
-            if (hasNext == false) throw NoSuchElementException()
-            hasNext = null
-            return ResultRow.create(rs, fieldsIndex)
+        init {
+            hasNext = rs.next()
+            if (hasNext) trackResultSet(transaction)
         }
 
-        override fun hasNext(): Boolean {
-            if (hasNext == null) hasNext = rs.next()
-            if (hasNext == false) {
-                rs.close()
-                untrackStatement(transaction, this)
-            }
-            return hasNext!!
+        override operator fun next(): ResultRow {
+            if (!hasNext) throw NoSuchElementException()
+            val result = ResultRow.create(rs, fieldsIndex)
+            hasNext = rs.next()
+            return result
         }
+
+        override fun hasNext(): Boolean = hasNext
     }
 
     companion object {
-        private val trackedOpenResultIterators = ConcurrentHashMap<Transaction, MutableSet<AbstractQuery<*>.ResultIterator>>()
-
-        private fun trackStatementOpen(transaction: Transaction, iterator: AbstractQuery<*>.ResultIterator) {
-            trackedOpenResultIterators.getOrPut(transaction) { ConcurrentHashMap.newKeySet() }.add(iterator)
-        }
-
-        private fun untrackStatement(transaction: Transaction, iterator: AbstractQuery<*>.ResultIterator) {
-            trackedOpenResultIterators[transaction]?.remove(iterator)
-        }
-
-        internal fun closeOpenedStatements(transaction: Transaction) {
-            trackedOpenResultIterators.remove(transaction)?.forEach {
-                if (!it.rs.isClosed) it.rs.close()
+        private fun trackResultSet(transaction: Transaction) {
+            val threshold = transaction.db.config.logTooMuchResultSetsThreshold
+            if (threshold > 0 && threshold < transaction.openResultSetsCount) {
+                val message =
+                    "Current opened result sets size ${transaction.openResultSetsCount} exceeds $threshold threshold for transaction ${transaction.id} "
+                val stackTrace = Exception(message).stackTraceToString()
+                exposedLogger.error(stackTrace)
             }
+            transaction.openResultSetsCount++
         }
     }
 }